fix(coding-agent): clear mcp chaos gate quality blockers
