+++ .omo/evidence/todo-21-quality-fix-verification-summary.md
@@ -1,0 +1,42 @@
+# TODO21 Quality Fix Verification Summary
+
+Worktree: `/Users/yeongyu/local-workspaces/senpi-wt/senpi-mcp-plugin-w2`
+Branch: `code-yeongyu/senpi-mcp-plugin-w2`
+Evidence root: `local-ignore/qa-evidence/20260707-mcp-w2-todo21-quality-fix/`
+
+Passed required commands:
+- `cd packages/coding-agent && npx tsx ../../node_modules/vitest/dist/cli.js --run test/suite/regressions/mcp-prefixed-extension-tool-allowlist.test.ts`
+  - PASS, 1 file / 1 test. Artifact: `final-regression-test.txt`.
+- `cd packages/coding-agent && npx tsx ../../node_modules/vitest/dist/cli.js --run test/mcp/connection.test.ts test/mcp/reconnect.test.ts test/mcp/ping-on-call.test.ts test/mcp/idle.test.ts test/mcp/service-lifecycle.test.ts test/mcp/register-call.test.ts test/mcp/commands.test.ts`
+  - PASS, 7 files / 51 tests. Artifact: `final-impacted-mcp-tests.txt`.
+- Gate reviewer no-excuse command exactly as supplied.
+  - PASS, no violations in 10 files. Artifact: `final-no-excuse-gate-command.txt`.
+- `npm run check`
+  - Initial post-split run failed on a missing type import; after fix PASS. Artifact: `rerun-npm-run-check-after-compile-fix.txt`.
+- `node .agents/skills/senpi-qa/scripts/mock-loop.mjs --self-test`
+  - PASS, 5/5, only localhost fake providers hit, real auth unchanged. Artifact: `final-senpi-qa-mock-loop-self-test.txt`.
+
+Pure LOC:
+- `connection.ts`: 245
+- `connection-types.ts`: 31
+- `mcp-prefixed-extension-tool-allowlist.test.ts`: 71
+- All gate-review files measured in `final-pure-loc-scope.txt` are <= 250.
+
+Chaos driver:
+- Invocation: `node local-ignore/qa-evidence/20260707-mcp-w2-todo21/todo21-chaos-driver.mjs --evidence-root local-ignore/qa-evidence/20260707-mcp-w2-todo21-quality-fix/chaos-current-diff`
+- PASS for scenarios 1-9, including the 30-minute keepalive soak.
+- Step 10 broad `npm test` failed; `npm run check` inside the driver passed.
+- Follow-up isolated reruns passed:
+  - `test/mcp/commands.test.ts test/mcp/startup-race.test.ts`
+  - `test/mcp/catalog-cache.test.ts`
+  - `test/mcp/catalog-cache.test.ts test/mcp/commands.test.ts test/mcp/startup-race.test.ts`
+  - `test/footer-data-provider.test.ts`
+
+Cleanup:
+- `final-cleanup-receipts.txt`: no fixture/model processes; no QA-created tmux sessions; chaos ports free.
+- Existing tmux session `ulw-dr` was present and not created by this task.
+- `chaos-current-diff/cleanup-receipt.txt`: `authUnchanged=true`, `processSweep=none`, ports free.
+
+Conclusion:
+- The three TODO21 final gate blockers are fixed.
+- The broad root `npm test` instability is recorded as residual risk and is outside this scoped quality fix.
